examples/notify/requirements.txt
bleak>=0.21.0
sounddevice
openai-whisper
